+++ texar/torch/modules/pretrained/__init__.py
@@ -19,6 +19,5 @@ from texar.torch.modules.pretrained.pretrained_base import *
 from texar.torch.modules.pretrained.bert import *
 from texar.torch.modules.pretrained.gpt2 import *
 from texar.torch.modules.pretrained.roberta import *
-from texar.torch.modules.pretrained.spanbert import *
 from texar.torch.modules.pretrained.xlnet import *
 from texar.torch.modules.pretrained.t5 import *
